xen/arm: raise MAX_VIRT_CPUS to 1024 on arm64
With a second redistributor region described by Xen and by the toolstack,
and with struct domain no longer growing with NR_CPUS, nothing structural
holds a guest to 128 vCPUs any more. Raise the limit to 1024.

Not to 4096, which is what the vMPIDR mapping could express and what
vgic_max_vcpus() already returns for a vGICv3. The binding constraint is
guest address space: at 128K per redistributor, 4096 of them need 512M, and
the guest layout has no 512M hole. GICR0 is followed immediately by the
vITS, the largest gap below the RAM banks is the ~191M between the vITS and
the vPCI ECAM window, and 4G to 8G is vPCI prefetch memory with RAM1
beginning at 8G. Reaching 4096 means either rearranging the layout, which
breaks every guest rather than only those wanting more than 128 vCPUs, or
putting redistributors above 4G where a 32-bit guest cannot follow. 1024
needs 128M and fits in the gap that exists.

The vCPU IDs stored by the vGIC have to be widened to match. pending_irq's
lpi_vcpu_id and vgic_irq_rank's vcpu[] are both uint8_t, which cannot hold
a vCPU ID past 255, and the BUILD_BUG_ONs already guarding them against
MAX_VIRT_CPUS are what would catch this. Both are accessed with
read_atomic()/write_atomic(), which handle a 2-byte type. The ITS collection
table entry is already uint16_t and needs nothing.

A vGICv2 guest is unaffected: ITARGETSR is an 8-bit CPU target mask, so
vgic_max_vcpus() returns 8 for it and that is what such a domain gets.

GUEST_MAX_VCPUS follows, being the toolstack's view of the same limit. It
sizes the static DSDT that mk_dsdt generates for ACPI guests, so that table
grows; each processor object's _STA is gated on the runtime \_SB.NCPU, so a
smaller domain still reports only the CPUs it has.

+++ xen/arch/arm/include/asm/vgic.h
@@ -84,7 +84,7 @@ struct pending_irq
     uint8_t lr;
     uint8_t priority;
     uint8_t lpi_priority;       /* Caches the priority if this is an LPI. */
-    uint8_t lpi_vcpu_id;        /* The VCPU for an LPI. */
+    uint16_t lpi_vcpu_id;       /* The VCPU for an LPI. */
     /* inflight is used to append instances of pending_irq to
      * vgic.inflight_irqs */
     struct list_head inflight;
@@ -124,7 +124,7 @@ struct vgic_irq_rank {
      * Use atomic operations to read/write the vcpu fields to avoid
      * taking the rank lock.
      */
-    uint8_t vcpu[32];
+    uint16_t vcpu[32];
 };
 
 struct vgic_dist {
